feat: divide main screen from main file
